projects/weather-bot/requirements.txt
py-clob-client>=0.17.0
requests>=2.31.0
scipy>=1.11.0
python-dotenv>=1.0.0
eth-account>=0.11.0
